+++ Tools/clinic/libclinic/__init__.py
@@ -32,6 +32,8 @@ from .utils import (
     Sentinels,
     unspecified,
     unknown,
+    Null,
+    NULL,
 )
 
 
@@ -68,6 +70,8 @@ __all__ = [
     "Sentinels",
     "unspecified",
     "unknown",
+    "Null",
+    "NULL",
 ]
