+++ Tools/clinic/libclinic/function.py
@@ -6,8 +6,9 @@ import functools
 import inspect
 from typing import Final, Any, TYPE_CHECKING
 if TYPE_CHECKING:
-    from clinic import Clinic, CReturnConverter, self_converter
+    from clinic import Clinic, CReturnConverter
     from libclinic.converter import CConverter
+    from libclinic.converters import self_converter
 
 from libclinic import VersionTuple, unspecified
